funcionando melo

+++ GoldenFeets/src/main/java/com/GoldenFeet/GoldenFeets/controller/ApiController.java
@@ -9,9 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
-import java.util.List; // <-- Importación necesaria para el nuevo método
+import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api")
@@ -22,40 +21,38 @@ public class ApiController {
 
     /**
      * Obtiene los detalles de un producto por su ID.
-     * Usado para la "Vista Rápida" en el catálogo.
      */
     @GetMapping("/productos/{id}")
-    public ResponseEntity<ProductoDTO> obtenerProducto(@PathVariable Long id) {
+    // --- CORRECCIÓN: Cambiado de Long a Integer ---
+    public ResponseEntity<ProductoDTO> obtenerProducto(@PathVariable Integer id) {
         return productoService.buscarPorId(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
-    // --- ENDPOINT AÑADIDO PARA LA LISTA DE DESEOS ---
     /**
      * Recibe una lista de IDs de productos y devuelve los detalles de cada uno.
-     * Esencial para que la página de Lista de Deseos pueda mostrar los productos.
      */
     @PostMapping("/productos/by-ids")
-    public ResponseEntity<List<ProductoDTO>> obtenerProductosPorIds(@RequestBody List<Long> ids) {
+    // --- CORRECCIÓN: Cambiado de List<Long> a List<Integer> ---
+    public ResponseEntity<List<ProductoDTO>> obtenerProductosPorIds(@RequestBody List<Integer> ids) {
         List<ProductoDTO> productos = productoService.listarPorIds(ids);
         return ResponseEntity.ok(productos);
     }
-    // --- FIN DEL CÓDIGO AÑADIDO ---
-
 
     /**
      * Agrega un producto al carrito de compras almacenado en la sesión.
      */
     @PostMapping("/carrito/agregar")
     public Map<String, Object> agregarAlCarrito(@RequestBody ItemCarritoRequest request, HttpSession session) {
+        // --- CORRECCIÓN: El mapa del carrito ahora usa Integer como clave ---
         @SuppressWarnings("unchecked")
-        Map<Long, Integer> carrito = (Map<Long, Integer>) session.getAttribute("carrito");
+        Map<Integer, Integer> carrito = (Map<Integer, Integer>) session.getAttribute("carrito");
         if (carrito == null) {
             carrito = new HashMap<>();
         }
 
-        long productoId = request.getProductoId();
+        int productoId = request.getProductoId();
         int cantidadActual = carrito.getOrDefault(productoId, 0);
         carrito.put(productoId, cantidadActual + request.getCantidad());
 
@@ -77,7 +74,8 @@ public class ApiController {
     @GetMapping("/carrito/total")
     public Map<String, Object> obtenerTotalCarrito(HttpSession session) {
         @SuppressWarnings("unchecked")
-        Map<Long, Integer> carrito = (Map<Long, Integer>) session.getAttribute("carrito");
+        // --- CORRECCIÓN: El mapa del carrito ahora usa Integer como clave ---
+        Map<Integer, Integer> carrito = (Map<Integer, Integer>) session.getAttribute("carrito");
 
         int totalItems = 0;
         if (carrito != null) {
@@ -95,6 +93,7 @@ public class ApiController {
  */
 @Data
 class ItemCarritoRequest {
-    private long productoId;
+    // --- CORRECCIÓN: Cambiado de long a int ---
+    private int productoId;
     private int cantidad;
 }
