+++ GoldenFeets/src/main/java/com/GoldenFeet/GoldenFeets/controller/CarritoController.java
@@ -24,28 +24,29 @@ public class CarritoController {
 
     @GetMapping("/carrito")
     public String verCarrito(HttpSession session, Model model) {
+        // --- CORRECCIÓN: El mapa del carrito usa Integer como clave ---
         @SuppressWarnings("unchecked")
-        Map<Long, Integer> carritoMap = (Map<Long, Integer>) session.getAttribute("carrito");
+        Map<Integer, Integer> carritoMap = (Map<Integer, Integer>) session.getAttribute("carrito");
 
         List<CarritoItemDTO> itemsDelCarrito = new ArrayList<>();
         BigDecimal subtotal = BigDecimal.ZERO;
 
         if (carritoMap != null && !carritoMap.isEmpty()) {
-            // 1. Obtenemos todos los IDs de los productos del carrito.
-            List<Long> productoIds = new ArrayList<>(carritoMap.keySet());
+            // 1. Obtenemos los IDs como una lista de Integer.
+            List<Integer> productoIds = new ArrayList<>(carritoMap.keySet());
 
-            // 2. Hacemos UNA SOLA CONSULTA a la base de datos para traer todos los productos.
+            // 2. Esta llamada ahora es correcta porque el parámetro es List<Integer>.
             List<ProductoDTO> productosEncontrados = productoService.listarPorIds(productoIds);
 
-            // 3. Convertimos la lista de productos en un mapa para un acceso rápido por ID.
-            Map<Long, ProductoDTO> productosMap = productosEncontrados.stream()
+            // 3. El mapa de productos ahora usa Integer como clave.
+            Map<Integer, ProductoDTO> productosMap = productosEncontrados.stream()
                     .collect(Collectors.toMap(ProductoDTO::id, Function.identity()));
 
-            // 4. Ahora recorremos el carrito original que tiene las cantidades.
-            for (Map.Entry<Long, Integer> entry : carritoMap.entrySet()) {
-                Long productoId = entry.getKey();
+            // 4. Recorremos el carrito original.
+            for (Map.Entry<Integer, Integer> entry : carritoMap.entrySet()) {
+                Integer productoId = entry.getKey(); // El tipo de la clave es Integer
                 Integer cantidad = entry.getValue();
-                ProductoDTO producto = productosMap.get(productoId); // Obtenemos el producto del mapa (muy rápido)
+                ProductoDTO producto = productosMap.get(productoId);
 
                 if (producto != null) {
                     BigDecimal precioItem = producto.precio().multiply(new BigDecimal(cantidad));
@@ -57,7 +58,7 @@ public class CarritoController {
 
         model.addAttribute("itemsCarrito", itemsDelCarrito);
         model.addAttribute("subtotal", subtotal);
-        model.addAttribute("total", subtotal); // Por ahora, el total es igual al subtotal
+        model.addAttribute("total", subtotal);
 
         return "carrito";
     }
