+++ GoldenFeets/src/main/java/com/GoldenFeet/GoldenFeets/service/ProductoService.java
@@ -2,9 +2,6 @@ package com.GoldenFeet.GoldenFeets.service;
 
 import com.GoldenFeet.GoldenFeets.dto.CategoriaDTO;
 import com.GoldenFeet.GoldenFeets.dto.ProductoDTO;
-// Categoria entity is not needed here as it's an implementation detail
-// import com.GoldenFeet.GoldenFeets.entity.Categoria;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -12,12 +9,10 @@ public interface ProductoService {
 
     List<ProductoDTO> listarTodos();
 
-    Optional<ProductoDTO> buscarPorId(Long id);
+    // --- CORRECCIÓN: Cambiado de Long a Integer ---
+    Optional<ProductoDTO> buscarPorId(Integer id);
 
     List<CategoriaDTO> listarCategorias();
-
-    // --- ESTA LÍNEA FUE ELIMINADA ---
-    // CategoriaDTO convertirACategoriaDTO(Categoria categoria);
 
     List<ProductoDTO> listarDestacados();
 
@@ -29,7 +24,8 @@ public interface ProductoService {
 
     List<String> listarMarcasDistintas();
 
-    List<ProductoDTO> listarPorIds(List<Long> ids);
+    // --- CORRECCIÓN: Cambiado de List<Long> a List<Integer> ---
+    List<ProductoDTO> listarPorIds(List<Integer> ids);
 
-    List<ProductoDTO> obtenerProductosRecientes(int cantidad); // Renamed 'limit' to 'cantidad' for consistency
+    List<ProductoDTO> obtenerProductosRecientes(int cantidad);
 }
